refactor: move all components into folders with their respective stories
