+++ src/pages/index.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Button from "@components/atoms/Button";
+import Button from "@components/atoms/Button/Button";
 import { styled } from "@theme";
 
 const Home: React.FC = () => {
